requirements.txt
runpod
torch
torchvision
ultralytics
timm
transformers
pillow
numpy